+++ authlete/django/handler/userinfo_request_handler.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2024 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -46,8 +46,8 @@ class UserInfoRequestHandler(BaseRequestHandler):
     def handle(self, request):
         """Handle a userinfo request.
 
-        This method calls Authlete's /api/auth/userinfo API and conditionally
-        /api/auth/userinfo/issue API.
+        This method calls Authlete's /auth/userinfo API and conditionally
+        /auth/userinfo/issue API.
 
         Args:
             request (django.http.HttpRequest)
@@ -59,8 +59,8 @@ class UserInfoRequestHandler(BaseRequestHandler):
             authlete.api.AuthleteApiException
         """
 
-        # Extract the access token from 'Authorization: Bearer {accessToken}'.
-        accessToken = RequestUtility.extractBearerToken(request)
+        # Extract the access token from the request.
+        accessToken = RequestUtility.extractAccessToken(request)
 
         if accessToken is None:
             # 400 Bad Request with a WWW-Authenticate header.
@@ -68,7 +68,7 @@ class UserInfoRequestHandler(BaseRequestHandler):
                 'Bearer error="invalid_token",error_description="An access token is required."')
 
         # Call Authlete's /api/auth/userinfo API.
-        res = self.__callUserInfoApi(accessToken)
+        res = self.__callUserInfoApi(request, accessToken)
 
         # 'action' in the response denotes the next action which the
         # implementation of the userinfo endpoint should take.
@@ -77,37 +77,54 @@ class UserInfoRequestHandler(BaseRequestHandler):
         # The content of the response to the client application.
         content = res.responseContent
 
+        # Additional HTTP headers.
+        headers = self.__prepareHeaders(res)
+
         if action == UserInfoAction.INTERNAL_SERVER_ERROR:
             # 500 Internal Server Error
-            return ResponseUtility.wwwAuthenticate(500, content)
+            return ResponseUtility.wwwAuthenticate(500, content, headers)
         elif action == UserInfoAction.BAD_REQUEST:
             # 400 Bad Request
-            return ResponseUtility.wwwAuthenticate(400, content)
+            return ResponseUtility.wwwAuthenticate(400, content, headers)
         elif action == UserInfoAction.UNAUTHORIZED:
             # 401 Unauthorized
-            return ResponseUtility.wwwAuthenticate(401, content)
+            return ResponseUtility.wwwAuthenticate(401, content, headers)
         elif action == UserInfoAction.FORBIDDEN:
             # 403 Forbidden
-            return ResponseUtility.wwwAuthenticate(403, content)
+            return ResponseUtility.wwwAuthenticate(403, content, headers)
         elif action == UserInfoAction.OK:
             # Return the user information.
-            return self.__getUserInfo(res)
+            return self.__getUserInfo(res, headers)
         else:
             # 500 Internal Server Error
-            # /api/auth/userinfo API returns an unknown action.
-            return self.unknownAction('/api/auth/userinfo')
+            # /auth/userinfo API returns an unknown action.
+            return self.unknownAction('/auth/userinfo')
 
 
-    def __callUserInfoApi(self, accessToken):
-        # Create a request for /api/auth/userinfo API.
+    def __callUserInfoApi(self, request, accessToken):
         req = UserInfoRequest()
+
+        # The access token.
         req.token = accessToken
+
+        # The request may contain a client certificate.
+        req.clientCertificate = RequestUtility.extractClientCert(request)
+
+        # The request may contain a DPoP proof JWT.
+        req.dpop = request.headers.get('DPoP')
 
         # Call /api/auth/userinfo API.
         return self.api.userinfo(req)
 
 
-    def __getUserInfo(self, response):
+    def __prepareHeaders(self, res):
+        if res.dpopNonce is not None:
+            return { 'DPoP-Nonce': res.dpopNonce }
+
+        return None
+
+
+    def __getUserInfo(self, response, headers):
         # Collect information about the user.
         claims = ClaimCollector(
             response.subject, response.claims, None, self._spi).collect()
@@ -116,10 +133,10 @@ class UserInfoRequestHandler(BaseRequestHandler):
         sub = self._spi.getSub()
 
         # Generate a response from the userinfo endpoint.
-        return self.__userInfoIssue(response.token, claims, sub)
+        return self.__userInfoIssue(response.token, claims, sub, headers)
 
 
-    def __userInfoIssue(self, token, claims, sub):
+    def __userInfoIssue(self, token, claims, sub, headers):
         # Call Authlete's /api/auth/userinfo/issue API.
         res = self.__callUserInfoIssueApi(token, claims, sub)
 
@@ -132,26 +149,26 @@ class UserInfoRequestHandler(BaseRequestHandler):
 
         if action == UserInfoIssueAction.INTERNAL_SERVER_ERROR:
             # 500 Internal Server Error
-            return ResponseUtility.wwwAuthenticate(500, content)
+            return ResponseUtility.wwwAuthenticate(500, content, headers)
         elif action == UserInfoIssueAction.BAD_REQUEST:
             # 400 Bad Request
-            return ResponseUtility.wwwAuthenticate(400, content)
+            return ResponseUtility.wwwAuthenticate(400, content, headers)
         elif action == UserInfoIssueAction.UNAUTHORIZED:
             # 401 Unauthorized
-            return ResponseUtility.wwwAuthenticate(401, content)
+            return ResponseUtility.wwwAuthenticate(401, content, headers)
         elif action == UserInfoIssueAction.FORBIDDEN:
             # 403 Forbidden
-            return ResponseUtility.wwwAuthenticate(403, content)
+            return ResponseUtility.wwwAuthenticate(403, content, headers)
         elif action == UserInfoIssueAction.JSON:
             # 200 OK, application/json; charset=UTF-8
-            return ResponseUtility.okJson(content)
+            return ResponseUtility.okJson(content, headers)
         elif action == UserInfoIssueAction.JWT:
             # 200 OK, application/jwt
-            return ResponseUtility.okJwt(content)
+            return ResponseUtility.okJwt(content, headers)
         else:
             # 500 Internal Server Error
-            # /api/auth/userinfo/issue API returns an unknown action.
-            return self.unknownAction('/api/auth/userinfo/issue')
+            # /auth/userinfo/issue API returns an unknown action.
+            return self.unknownAction('/auth/userinfo/issue')
 
 
     def __callUserInfoIssueApi(self, token, claims, sub):